backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
python-multipart
httpx
python-dotenv
